+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 ### Added
 * Show Flaw Labels on Flaw List Component (`OSIDB-3805`)
+* Show Flaw contributors table on edit view (`OSIDB-3806`)
 
 ### Fixed
 * Fix incorrect affects ordering by Impact and Resolution (`OSIDB-3480`)